# server/docs/swagger.yaml
basePath: /api
definitions:
  exercise.CreateExerciseRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  exercise.CreateExerciseResponse:
    properties:
      created_at:
        example: "2023-01-01T15:04:05Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Bench Press
        type: string
      updated_at:
        example: "2023-01-01T15:04:05Z"
        type: string
      user_id:
        example: user-123
        type: string
    required:
    - created_at
    - id
    - name
    - updated_at
    - user_id
    type: object
...
  exercise.ExerciseResponse:
    properties:
      created_at:
        example: "2023-01-01T15:04:05Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Bench Press
        type: string
      updated_at:
        example: "2023-01-01T15:04:05Z"
        type: string
      user_id:
        example: user-123
        type: string
    required:
    - created_at
    - id
    - name
    - updated_at
    - user_id
    type: object
  exercise.ExerciseWithSetsResponse:
    properties:
      exercise_id:
        example: 1
        type: integer
      exercise_name:
        example: Bench Press
        type: string
      exercise_order:
        example: 0
        type: integer
      reps:
        example: 10
        type: integer
      set_id:
        example: 1
        type: integer
      set_order:
        example: 1
        type: integer
      set_type:
        example: working
        type: string
      volume:
        example: 2250.5
        type: number
      weight:
        example: 225.5
        type: number
      workout_date:
        example: "2023-01-01T15:04:05Z"
        type: string
      workout_id:
        example: 1
        type: integer
      workout_notes:
        example: Great workout today
        type: string
    required:
    - exercise_id
    - exercise_name
    - reps
    - set_id
    - set_type
    - volume
    - workout_date
    - workout_id
    type: object
...
  exercise.RecentSetsResponse:
    properties:
      created_at:
        example: "2023-01-01T15:04:05Z"
        type: string
      exercise_order:
        example: 0
        type: integer
      reps:
        example: 10
        type: integer
      set_id:
        example: 1
        type: integer
      set_order:
        example: 2
        type: integer
      weight:
        example: 225.5
        type: number
      workout_date:
        example: "2023-01-01T15:04:05Z"
        type: string
      workout_id:
        example: 1
        type: integer
    required:
    - created_at
    - reps
    - set_id
    - workout_date
    - workout_id
    type: object
...
  exercise.UpdateExerciseNameRequest:
    properties:
      name:
        maxLength: 256
        type: string
    required:
    - name
    type: object
...
  response.ErrorResponse:
    properties:
      message:
        example: Error message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A fitness tracking application API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FitTrack API
  version: "1.0"
paths:
  /exercises:
    get:
      consumes:
      - application/json
      description: Get all exercises for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.ExerciseResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: List exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      description: Get an existing exercise by name or create it if it doesn't exist
      parameters:
      - description: Exercise data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/exercise.CreateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.CreateExerciseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Get or create exercise
      tags:
      - exercises
  /exercises/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific exercise and all its associated sets. Only the
        owner of the exercise can delete it.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Exercise deleted successfully
        "400":
          description: Bad Request - Invalid exercise ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized - Invalid token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found - Exercise not found or doesn't belong to user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Delete an exercise
      tags:
      - exercises
    get:
      consumes:
      - application/json
      description: Get a specific exercise with all its sets from workouts. Returns
        empty array when exercise exists but has no sets.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success (may be empty array if exercise has no sets)
          schema:
            items:
              $ref: '#/definitions/exercise.ExerciseWithSetsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found - Exercise not found or doesn't belong to user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Get exercise with sets
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      description: Update the name of an exercise by ID (must belong to authenticated
        user)
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise name update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/exercise.UpdateExerciseNameRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Exercise name updated successfully
        "400":
          description: Bad Request - Invalid exercise ID or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized - Invalid token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found - Exercise not found or doesn't belong to user
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict - Exercise name already exists
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Update an exercise name
      tags:
      - exercises
...
  /exercises/{id}/recent-sets:
    get:
      consumes:
      - application/json
      description: Get the 3 most recent sets for a specific exercise. Returns empty
        array when exercise has no sets.
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success (may be empty array)
          schema:
            items:
              $ref: '#/definitions/exercise.RecentSetsResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - StackAuth: []
      summary: Get recent sets for exercise
      tags:
      - exercises
...
schemes:
- http
securityDefinitions:
  StackAuth:
    in: header
    name: x-stack-access-token
    type: apiKey
swagger: "2.0"